logic23344423edсxe

+++ bot.js
@@ -1007,8 +1007,8 @@ bot.on("message", async (msg) => {
         },
       }
     );
-    console.log(user.message_to_delete, "message_to_delete");
-    if (user.message_to_delete) {
+
+    if (user && user.message_to_delete) {
       await bot.deleteMessage(chatId, user.message_to_delete);
     }
     await collectionUser.updateOne(
